split parser_impl in two files
