+++ app/src/parser_impl.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
-*  (c) 2019 Zondax GmbH
+*  (c) 2019 - 2022  Zondax GmbH
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -174,7 +174,13 @@ parser_error_t _readEra(parser_context_t *c, pd_ExtrinsicEra_t *v);
 
 parser_error_t _readTx(parser_context_t *c, parser_tx_t *v);
 
+parser_error_t _checkVersions(parser_context_t *c);
+
 uint16_t _getAddressType();
+
+parser_error_t _readCompactIndex(parser_context_t *c, pd_CompactIndex_t *v);
+
+uint16_t _detectAddressType(const parser_context_t *c);
 
 parser_error_t _toStringCompactInt(const compactInt_t *c, uint8_t decimalPlaces,
                                    char postfix[],
